Update ModelMakeCommand.php
When the model name is in all uppercase, the table name in the migration file needs to be updated to match the expected naming convention.

For example, if the model name is SOA, the old table name might be s_o_as. To fix this, you should rename the table to s_o_a_s, which follows Laravel's default table naming convention for uppercase model names.